LayoutTests/accessibility/aria-help-expected.txt

This tests that aria-help attribute works as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS platformValueForW3CName(axButton) is "title"
PASS platformValueForW3CDescription(axButton) is "click here"
PASS successfullyParsed is true

TEST COMPLETE
